Adapted NewUI to Trimpz. Added enhancements(see description).
Save settings only when they change, not in main loop.
Add default values to booleans.
Changed default values for value settings to use value instead of a string.
Changed ".enabled" for booleans to ".value" to fit in with other settings.  There's already a ".type" that can be used if needed.
GetPageSetting() is no longer necessary in main code.
Arranged variables according to use instead of type.